requirements.txt
bcrypt==2.0.0
cffi==1.5.2
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.2
djangorestframework==3.3.2
psycopg2==2.6.1
pycparser==2.14
six==1.10.0
static3==0.6.1
waitress==0.8.10
wheel==0.24.0
whitenoise==2.0.6
